+++ src/components/Mypage/ScrapListBox.tsx
